Refactored IncrementalCachingTests
